fix a bit

+++ detection/predict.py
@@ -63,8 +63,8 @@ def crop_box(img, boxes, image_name, out_folder, num_boxes=0, save_csv=True):
         max_y = min(h, max(y1,y2,y3,y4))
         
         if num_boxes==0:
-            tw = np.sqrt((x1-x2)**2 + (y1-y2)**2)
-            th = np.sqrt((x1-x4)**2 + (y1-y4)**2)
+            tw = int(np.sqrt((x1-x2)**2 + (y1-y2)**2))
+            th = int(np.sqrt((x1-x4)**2 + (y1-y4)**2))
             pt1 = np.float32([(x1,y1),(x2,y2),(x3,y3),(x4,y4)])
             pt2 = np.float32([[0, 0],
                               [tw - 1, 0],
